refactor(terminals): share shell availability hook
