+++ src/renderer/lib/hooks/use-terminal-shell-availability.ts
